updates data imports to make run-seed file and test file more readable

+++ prisma/seed/seed.ts
@@ -13,12 +13,12 @@ const client = new PrismaClient();
 async function deleteTables() {
   const tableNames: Array<{ tablename: String }> =
     await client.$queryRaw`SELECT tablename FROM pg_tables WHERE schemaname='public'`;
-
   const tables = tableNames
     .map(({ tablename }) => tablename)
     .map((name) => `"public"."${name}"`)
     .join(", ");
   try {
+    if (tableNames.length === 0) return;
     await client.$executeRawUnsafe(
       `TRUNCATE TABLE ${tables} RESTART IDENTITY;`
     );
